Additional Cloudaux resources for GCP.
Fitbit has been using Security Monkey internally to monitor our GCP
environments and we'd like to start contributing back some of that code
to the main Security Monkey project.

This PR contains the following:
 * Support for GCE Instances
 * Support for GCE Persistent Disks
 * Support for GCE Addresses (global and regional)
 * Support for GCE Forwarding Rules (global and regional)
 * Support for CRM (i.e., Google Projects and their related IAM bindings)

I've also modified the `tox.ini` to enable the GCP tests which were
previously added but not hooked up to the CI.  In the process I've fixed
a broken GCP test.

This is a pre-requisite for Security Monkey watcher and auditor support
for these resource types.